Fil de renderitzat + update món a temps fixat Variable TPS a main per controlar quantes vegades s'ha d'actualitzar el món, per segon Vsync activat: glSwapInteval(1) Els esdeveniments se segueixen processant quan es pot. Zones d'exclusió mútua: llista de chunks visibles del món i llista de cares transparents a cada chunk

+++ CPlusRaft/Blocks/Block.cpp
@@ -16,7 +16,7 @@ Block::Block() {
 	//this->parent = 0;
 }
 
-void Block::update(int delta, World* world) {
+void Block::update(float delta, World* world) {
 
 }
