+++ CPlusRaft/Blocks/LiquidBlock.h
@@ -23,7 +23,7 @@ public:
 
 	void interact(World* world);
 
-	void update(int delta, World* world);
+	void update(float delta, World* world);
 
 	void setLvl(int lvl);
